refactor: Divide `github_actions` into modules
